refactor: Flatten directory structure
- Each notification view has a directory with a template and sample data
