+++ src/ducks/notifications/HealthBillLinked/index.js
@@ -1,12 +1,12 @@
-import NotificationView from './BaseNotificationView'
-import { isHealthExpense } from '../categories/helpers'
+import NotificationView from 'ducks/notifications/BaseNotificationView'
+import { isHealthExpense } from 'ducks/categories/helpers'
 import { Bill } from 'models'
 import keyBy from 'lodash/keyBy'
 import log from 'cozy-logger'
-import { treatedByFormat } from './html/utils'
-import { getReimbursementBillIds } from './helpers'
-import { prepareTransactions, getCurrentDate } from './html/utils'
-import template from './html/templates/health-bill-linked.hbs'
+import { treatedByFormat } from 'ducks/notifications/utils'
+import { getReimbursementBillIds } from 'ducks/notifications/helpers'
+import { prepareTransactions, getCurrentDate } from 'ducks/notifications/utils'
+import template from './template.hbs'
 import { toText } from 'cozy-notifications'
 
 const ACCOUNT_SEL = '.js-account'
