+++ src/ducks/notifications/TransactionGreater/template.hbs
