+++ src/ducks/notifications/assets/icons/icon-cat-activities.png
