+++ src/ducks/notifications/assets/icons/icon-cat-dailyLife.png
