+++ src/ducks/notifications/assets/icons/icon-cat-educationAndTraining.png
