+++ src/ducks/notifications/assets/icons/icon-cat-excludeFromBudgetCat.png
