+++ src/ducks/notifications/assets/icons/icon-cat-goingOutAndTravel.png
