+++ src/ducks/notifications/assets/icons/icon-cat-health.png
