+++ src/ducks/notifications/assets/icons/icon-cat-incomeCat.png
