+++ src/ducks/notifications/assets/icons/icon-cat-tax.png
