+++ src/ducks/notifications/assets/icons/icon-cat-transportation.png
