+++ src/ducks/notifications/assets/icons/icon-cat-uncategorized.png
