+++ src/ducks/notifications/assets/icons/icon-hourglass-blue.png
