+++ src/ducks/notifications/assets/icons/icon-hourglass-red.png
