+++ src/ducks/notifications/assets/logo-bank.png
