+++ src/ducks/notifications/assets/logo-cozy.png
