+++ src/ducks/notifications/bank-layout.hbs
